% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_variable_generator.R
\name{decode}
\alias{decode}
\title{Decode a binary message.}
\usage{
decode(h_tree, encoded_msg)
}
\arguments{
\item{h_tree}{Huffman root tree.}

\item{encoded_msg}{Binary vector.}
}
\value{
Message decoded through symbols.
}
\description{
Decode a binary message.
}
