% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{.add_dyadic_decomposition}
\alias{.add_dyadic_decomposition}
\title{Add all dyadic atom of a frequency into a priority queue.}
\usage{
.add_dyadic_decomposition(pq, freq, symbol)
}
\arguments{
\item{pq}{The priority queue.}

\item{freq}{The frequency to split.}

\item{symbol}{The symbol associate with.}
}
\value{
The priority queue.
}
\description{
Add all dyadic atom of a frequency into a priority queue.
}
