% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_variable_generator.R
\name{.get_huffman_tree}
\alias{.get_huffman_tree}
\title{Construct the Huffman tree from frequencies.}
\usage{
.get_huffman_tree(nodes_pq)
}
\arguments{
\item{nodes_pq}{The priority queue containing all dyadic frequencies.}
}
\value{
The root of the Huffman tree.
}
\description{
Construct the Huffman tree from frequencies.
}
