% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{generate_discrete_dist}
\alias{generate_discrete_dist}
\title{Generate a discrete distribution (based on exponential v.a).}
\usage{
generate_discrete_dist(size)
}
\arguments{
\item{size}{Number of value.}
}
\value{
A numeric vector.
}
\description{
Generate a discrete distribution (based on exponential v.a).
}
