% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_variable_generator.R
\name{generate_va}
\alias{generate_va}
\title{Generate a v.a from frequencies.}
\usage{
generate_va(freqs)
}
\arguments{
\item{freqs}{The frequencies vector.}
}
\value{
The root of the Huffman tree.
}
\description{
Generate a v.a from frequencies.
}
