% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{get_entropy}
\alias{get_entropy}
\title{Calculate the entropy of an v.a.}
\usage{
get_entropy(freqs)
}
\arguments{
\item{freqs}{Frequencies vector.}
}
\value{
The entropy.
}
\description{
Calculate the entropy of an v.a.
}
