% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{get_nodes}
\alias{get_nodes}
\title{Map a vector of frequencies to priority queue of dyadic nodes.}
\usage{
get_nodes(frequencies)
}
\arguments{
\item{frequencies}{Frequencies vector.}
}
\value{
A priority queue of nodes.
}
\description{
Map a vector of frequencies to priority queue of dyadic nodes.
}
