% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_tests.R
\name{tests_random_va}
\alias{tests_random_va}
\title{Test a simulation of a v.a}
\usage{
tests_random_va(freqs, plot = TRUE)
}
\arguments{
\item{freqs}{The frequencies vector.}

\item{plot}{If TRUE show the plot.}
}
\value{
Nothing.
}
\description{
Test a simulation of a v.a
}
